CIAPP-86 Add support for crashes
Add `PLCrashReporter` as crash handler, configure with partially symbolicated output
When a test starts it serializes its span in a simplified format ( `SimpleSpanData` ) and adds it as the `customData` of `PLCrashReporter`,
if during the test it crashes this spn is saved along the crash report.
When the test are going to start and the crash reporter is installed it check for existing crash report, if it exists,
then reconstructs the crashed test span and uploads the span.

Also change attribute name for "error.message" so it is the same used internally in Datadog