CIAPP-386 Capture stdout/stderr for Swift tests
Create `StderrCapture` and `StdoutCapture` classes for capturing and sending the strings to `DDTracer.logString()` variants
Implementation is different because we can capture stdout synchronously and add it to the active span in the test, for stderr an asynchronous method is needed.
Also add environment variables for disabling some of this functionality.