CIAPP-355 Add better support for network autoinstrumentation
Replace existing Network instrumentation classes with a new implementation based on swizzling `URLSession`  and `URLSessionDelegate` methods, in the spirit of PDAFNetworkDomainController (PonyDebugger) or FlexNetworkObserver. Most of the logic is implemented in a pair of classes `DDNetworkInstrumentation` which handles the method swizzling and calls `DDNetworkActivityLogger` which creates the spans and fill the span attributes with the data used in the network call.  It should support auto instrumenting `URLSession` or any other library based on it.

Expanding from previous implementation, by default it also logs request and response headers (filtering content of private or not configured headers),  and can be configured to log the first bytes of the payload (both request and response). Some documentation for configuring is added to the README

Some other general and small changes added, most important to `DDTracer.activeSpan`, that now is a method that returns the most appropiate span for using.