Automatically symbolicate crashes when running tests on macOS or with simulator When processing a crash report, check if it is macOS or simulator and call DDSymbolicator with the unsymbolicated crash Data.
DDSymbolicator locates all DSYM files in the build folder, and tries to desymbolicate line by line. If there is a DSYM file for the symbol it uses atos, if not, it prints the nearest run-time symbol.

It the stack is too big for apm intake (5000 characters) it uses dd.stack.01, dd.stack.02, ... to fill the complete crash report